refactor in mobile

+++ Mobile/DriveDrop.Core/DriveDrop.Core/ViewModels/LoginViewModel.cs
@@ -2,6 +2,7 @@
 using DriveDrop.Core.Models.User;
 using DriveDrop.Core.Services.Identity;
 using DriveDrop.Core.Services.OpenUrl;
+using DriveDrop.Core.Services.User;
 using DriveDrop.Core.Validations;
 using DriveDrop.Core.ViewModels.Base;
 using IdentityModel.Client;
@@ -24,13 +25,17 @@ namespace DriveDrop.Core.ViewModels
 
         private IOpenUrlService _openUrlService;
         private IIdentityService _identityService;
+        private IUserService _userService;
 
         public LoginViewModel(
             IOpenUrlService openUrlService,
-            IIdentityService identityService)
+            IIdentityService identityService,
+            IUserService userService)
         {
             _openUrlService = openUrlService;
             _identityService = identityService;
+
+            _userService = userService;
 
             _userName = new ValidatableObject<string>();
             _password = new ValidatableObject<string>();
@@ -117,6 +122,12 @@ namespace DriveDrop.Core.ViewModels
             }
         }
 
+        public ICommand DriverCommand => new Command(async () => await NewDriversAsync());
+
+        public ICommand SenderCommand => new Command(async () => await DialogService.ShowAlertAsync("Option no available, try again later.", "Oops!", "Ok"));
+
+
+
         public ICommand MockSignInCommand => new Command(async () => await MockSignInAsync());
 
         public ICommand SignInCommand => new Command(async () => await SignInAsync());
@@ -181,8 +192,43 @@ namespace DriveDrop.Core.ViewModels
 
             IsBusy = false;
         }
-
         private async Task SignInAsync()
+        {
+            IsBusy = true;
+            //IsValid = true;
+            bool isValid = Validate();
+            bool isAuthenticated = false;
+
+            if (isValid)
+            {
+                try
+                {
+                    await Task.Delay(500);
+                    LoginUrl = _identityService.CreateAuthorizationRequest();
+                  // IsLogin = true;
+                    //isAuthenticated = true;
+                }
+                catch (Exception ex)
+                {
+                    Debug.WriteLine($"[SignIn] Error signing in: {ex}");
+                }
+            }
+            else
+            {
+                IsValid = false;
+            }
+
+            if (isAuthenticated)
+            {
+                Settings.AuthAccessToken = GlobalSetting.Instance.AuthToken;
+
+                await NavigationService.NavigateToAsync<MainViewModel>();
+                await NavigationService.RemoveLastFromBackStackAsync();
+            }
+
+            IsBusy = false;
+        }
+        private async Task BackupSignInAsync()
         {
             IsBusy = true;
 
@@ -236,18 +282,35 @@ namespace DriveDrop.Core.ViewModels
                 var authResponse = new AuthorizeResponse(url);
                 if (!string.IsNullOrWhiteSpace(authResponse.Code))
                 {
+
                     var userToken = await _identityService.GetTokenAsync(authResponse.Code);
                     string accessToken = userToken.AccessToken;
 
-                    if (!string.IsNullOrWhiteSpace(accessToken))
+                    //valid user name and password
+                     
+                    var response = await _userService.GetUserInfoAsync(userToken.AccessToken, UserName.Value,Password.Value);
+
+                    if (response == "User Valid")
                     {
-                        Settings.AuthAccessToken = accessToken;
-                        Settings.AuthIdToken = authResponse.IdentityToken;
-                        await NavigationService.NavigateToAsync<MainViewModel>();
-                        await NavigationService.RemoveLastFromBackStackAsync();
+
+                        if (!string.IsNullOrWhiteSpace(accessToken))
+                        {
+                            Settings.AuthAccessToken = accessToken;
+                            Settings.AuthIdToken = authResponse.IdentityToken;
+
+
+                            await NavigationService.NavigateToAsync<MainViewModel>();
+                            await NavigationService.RemoveLastFromBackStackAsync();
+
+                        }
                     }
                 }
             }
+        }
+
+        private async Task NewDriversAsync()
+        {
+            await NavigationService.NavigateToAsync<NewDriverViewModel>();
         }
 
         private async Task SettingsAsync()
